Fix multiple setup calls in acceptance tests
Because the initializer reopens the Route class itself, if it runs multiple times, it will create a chain of setupController() super-class methods, one for each time it runs. So when running acceptance tests, each one will add a new sub-class method, meaning in each nth test the controller's setup() method will be called n times. The same is true of the reset() method, but the call order will be setup(), setup(), setup(), ..., reset(), reset(), reset(), violating the controller's expectations.